+++ src/main/java/lotto/domain/lottogenerator/LottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain.lotto.lottogenerator;
+package lotto.domain.lottogenerator;
 
 import java.util.List;
 import lotto.domain.lotto.LottoNumber;
