+++ src/test/java/lotto/domain/lottogenerator/RandomLottoGeneratorTest.java
@@ -1,7 +1,6 @@
-package lotto.domain.lotto.lottogenerator;
+package lotto.domain.lottogenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 import lotto.domain.lotto.LottoNumber;
